python-standalone/requirements.txt
gradio==4.12.0
transformers==4.36.0
torch==2.1.0
torchvision==0.16.0
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
